*** empty log message ***

+++ org/opensha/sha/earthquake/rupForecastImpl/MeanUCERF/MeanUCERF2.java
@@ -117,7 +117,7 @@ public class MeanUCERF2 extends EqkRupForecast {
 
 	// For rupture offset lenth along fault parameter
 	public final static String RUP_OFFSET_PARAM_NAME ="Rupture Offset";
-	private Double DEFAULT_RUP_OFFSET_VAL= new Double(10);
+	private Double DEFAULT_RUP_OFFSET_VAL= new Double(5);
 	private final static String RUP_OFFSET_PARAM_UNITS = "km";
 	private final static String RUP_OFFSET_PARAM_INFO = "Length of offset for floating ruptures";
 	public final static double RUP_OFFSET_PARAM_MIN = 1;
